+++ src/main/java/absyn/Exp.java
@@ -4,5 +4,5 @@ import java.util.Map;
 
 public abstract class Exp extends AST {
 
-   public abstract Integer eval(Map<String, Integer> mem);
+   public abstract Double eval(Map<String, Double> mem);
 }
